refactor: move types to common package
to reduce many areas of duplication
