+++ plugins/announcements-backend/src/service/persistence/CategoriesDatabase.test.ts
@@ -2,7 +2,7 @@ import { Knex } from 'knex';
 import { CategoriesDatabase } from './CategoriesDatabase';
 import { TestDatabases } from '@backstage/backend-test-utils';
 import { initializePersistenceContext } from './persistenceContext';
-import { Category } from '../model';
+import { Category } from '@procore-oss/backstage-plugin-announcements-common';
 
 function createDatabaseManager(client: Knex, skipMigrations: boolean = false) {
   return {
