+++ plugins/announcements-common/src/index.ts
@@ -1,1 +1,2 @@
 export * from './permissions';
+export * from './types';
