+++ plugins/announcements-common/src/types.ts
@@ -1,0 +1,19 @@
+export type Category = {
+  slug: string;
+  title: string;
+};
+
+export type Announcement = {
+  id: string;
+  category?: Category;
+  publisher: string;
+  title: string;
+  excerpt: string;
+  body: string;
+  created_at: string;
+};
+
+export type AnnouncementsList = {
+  count: number;
+  results: Announcement[];
+};
