+++ plugins/announcements-react/package.json
@@ -28,6 +28,7 @@
     "@backstage/core-plugin-api": "^1.8.2",
     "@backstage/errors": "^1.2.3",
     "@material-ui/core": "^4.12.2",
+    "@procore-oss/backstage-plugin-announcements-common": "workspace:^",
     "luxon": "^3.4.4"
   },
   "peerDependencies": {
