+++ plugins/announcements-react/src/apis/types.ts
@@ -1,22 +1,4 @@
-export type Category = {
-  slug: string;
-  title: string;
-};
-
-export type Announcement = {
-  id: string;
-  category?: Category;
-  publisher: string;
-  title: string;
-  excerpt: string;
-  body: string;
-  created_at: string;
-};
-
-export type AnnouncementsList = {
-  count: number;
-  results: Announcement[];
-};
+import { Announcement } from '@procore-oss/backstage-plugin-announcements-common';
 
 export type CreateAnnouncementRequest = Omit<
   Announcement,
