+++ plugins/announcements/src/components/AnnouncementsPage/useDeleteAnnouncementDialogState.tsx
@@ -1,5 +1,5 @@
+import { Announcement } from '@procore-oss/backstage-plugin-announcements-common';
 import { useCallback, useState } from 'react';
-import { Announcement } from '../../api';
 
 export type DeleteAnnouncementDialogState = {
   open: (a: Announcement) => void;
