+++ plugins/announcements/src/components/CreateAnnouncementPage/CreateAnnouncementPage.tsx
@@ -2,13 +2,13 @@ import React, { ReactNode } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { Page, Header, Content } from '@backstage/core-components';
 import { alertApiRef, useApi, useRouteRef } from '@backstage/core-plugin-api';
-import { Announcement } from '../../api';
 import { rootRouteRef } from '../../routes';
 import { AnnouncementForm } from '../AnnouncementForm';
 import {
   CreateAnnouncementRequest,
   announcementsApiRef,
 } from '@procore-oss/backstage-plugin-announcements-react';
+import { Announcement } from '@procore-oss/backstage-plugin-announcements-common';
 
 type CreateAnnouncementPageProps = {
   themeId: string;
